refactor to only need history drawer
